requirements.txt
asgiref==3.3.2
Django==3.2.10
django-crispy-forms==1.11.2
django-environ==0.4.5
django-extensions==3.1.1
django-filter==2.4.0
django-rest-passwordreset==1.2.1
django-webpack-loader==1.0.0
djangorestframework==3.12.4
drf-nested-routers==0.93.3
gunicorn==20.1.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
